[Refactor] Update files structure
