+++ src/drug/drug.ts
@@ -1,0 +1,11 @@
+export class Drug {
+    name: string;
+    expiresIn: number;
+    benefit: number;
+    constructor(name: string, expiresIn: number, benefit: number) {
+      this.name = name;
+      this.expiresIn = expiresIn;
+      this.benefit = benefit;
+    }
+  }
+
