+++ src/pharmacy/pharmacy.ts
@@ -1,13 +1,4 @@
-export class Drug {
-  name: string;
-  expiresIn: number;
-  benefit: number;
-  constructor(name: string, expiresIn: number, benefit: number) {
-    this.name = name;
-    this.expiresIn = expiresIn;
-    this.benefit = benefit;
-  }
-}
+import { Drug } from "../drug/drug";
 
 export class Pharmacy {
   drugs: Drug[];
